requirements.txt
google-genai
edge-tts
moviepy==1.0.3
openai-whisper
pydub
python-dotenv
mlx-whisper